add pip calculator
